orion-services/orion-insight-wrangler/requirements.txt
functions-framework==3.5.0
google-cloud-pubsub==2.22.0
google-cloud-firestore==2.16.0
google-cloud-aiplatform==1.71.1
trafilatura==1.9.0
spacy==3.7.5
numpy==1.26.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz
pytube==15.0.0
PyYAML==6.0.1